% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.min.prob.mean.R
\name{best.min.prob.mean}
\alias{best.min.prob.mean}
\title{Find mean expected SR over min probs}
\usage{
best.min.prob.mean(y, probs, n)
}
\arguments{
\item{y}{numeric/factor, 1 if up, -1 if down}

\item{probs}{numeric, probability of up}

\item{n}{number of stocks}
}
\value{
Named numeric containing the results of the best value for min.prob, as well as the mean sr, mean accuracy across min probs: c(min.prob, accuracy, freq, sr, mean.sr, mean.accuracy)
}
\description{
Find mean expected SR over min probs
}
