% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.forecasts.R
\name{combine.forecasts}
\alias{combine.forecasts}
\title{Combine Forecasts (WIP)
Combine forecasts of various classifiers into one single forecast}
\usage{
combine.forecasts(x, y)
}
\arguments{
\item{x:}{matrix where each line is an observation, and each colum contains the probability that the class is "positive "up" estimated by one model. Colnames are the name of the models.}

\item{y:}{vector of true classes}
}
\value{
list containing vector of combined probs and named numerics of coefficients
}
\description{
Combine Forecasts (WIP)
Combine forecasts of various classifiers into one single forecast
}
