% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.prices.direction.R
\name{compute.price.direction}
\alias{compute.price.direction}
\title{Compute Price Direction
Computes a variable describing the direction of future price changes}
\usage{
compute.price.direction(prices, daily.start = 1, horizon, min.return = 0,
  loading = T)
}
\arguments{
\item{prices}{xts where one numeric column has to be called "open" and one "close"}

\item{daily.start}{DEPRECATED (which observation number should be the first one for which the variable is computed each day; numeric)}

\item{horizon}{prediction horizon; numeric}

\item{min.return}{minimum absolute return in horizon to be classified as -1 or +1; default=0; numeric}

\item{loading}{logical telling if loading bar should be displayed}
}
\value{
xts with one numeric (values -1, 0, +1, NA) column and same indices as prices. last 'horizon' values of each day are NA
}
\description{
Compute Price Direction
Computes a variable describing the direction of future price changes
}
