% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.prices.direction.close.to.close.R
\name{compute.prices.direction.close.to.close}
\alias{compute.prices.direction.close.to.close}
\title{DEPRECATED: Compute Price Direction from close to close
Computes a variable describing the direction of future price changes}
\usage{
compute.prices.direction.close.to.close(prices, daily.start, horizon,
  min.return = 0, loading = T)
}
\arguments{
\item{prices}{xts where one numeric column has to be called "close"}

\item{daily.start}{which observation number should be the first one for which the variable is computed each day; numeric}

\item{horizon}{prediction horizon; numeric}

\item{min.return}{minimum absolute return in horizon to be classified as -1 or +1; default=0; numeric}

\item{loading}{logical telling if loading bar should be displayed}
}
\value{
xts with one numeric (values -1, 0, +1, NA) column and same indices as prices. last 'horizon' values of each day are NA
}
\description{
DEPRECATED: Compute Price Direction from close to close
Computes a variable describing the direction of future price changes
}
