% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.bound.R
\name{confidence.bound}
\alias{confidence.bound}
\title{Confidence Bound}
\usage{
confidence.bound(p, n, alpha = 0.05, lower = T)
}
\arguments{
\item{p}{Sample proportion; numeric}

\item{n}{Number of observations; numeric}

\item{alpha}{Unilateral probability of being outside of the bound; numeric}

\item{lower}{Lower or upper bound; logi}
}
\value{
value of the bound (numeric)
}
\description{
Confidence Bound
}
\examples{
confidence.bound(0.65, 110, 0.05, T)
}
