% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.min.prob.mean.R
\name{evaluate.model}
\alias{evaluate.model}
\title{Compute performance metrics over all min probs}
\usage{
evaluate.model(y, probs, n)
}
\arguments{
\item{y}{numeric/factor, 1 if up, -1 if down}

\item{probs}{numeric, probability of up}

\item{n}{number of stocks}
}
\value{
Named numeric containing the results of the best value for min.prop, as well as the mean sr across min probs: c(min.prob, accuracy, freq, sr, mean.sr)
}
\description{
Compute performance metrics over all min probs
}
