% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.predictions.num.R
\name{evaluate.predictions.num}
\alias{evaluate.predictions.num}
\title{Evaluate predictions performance against true label}
\usage{
evaluate.predictions.num(y, probs, min.prob, n, check = F)
}
\arguments{
\item{y}{: numeric (1 up, 0 down)}

\item{probs}{: numeric probability of up, second is down}

\item{n}{number of stocks}

\item{min.prob:}{numeric indicating minimum confidence required to make a prediction}

\item{check:}{logical indicating if dimension checks should be performed}
}
\value{
Named numeric containing c(min.prob, accuracy, freq, sr), or "no predictions" if min.prob is too high
}
\description{
Evaluate predictions performance against true label
}
