% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.predictions.R
\name{evaluate.predictions}
\alias{evaluate.predictions}
\title{Evaluate predictions performance against true label}
\usage{
evaluate.predictions(y, probs, min.prob, n = 7, check = F)
}
\arguments{
\item{y}{: numeric/factor, 1 if up, -1 if down}

\item{probs}{: numeric probability of up}

\item{n}{number of stocks (only relevant if sr is needed)}

\item{min.prob:}{numeric indicating minimum confidence required to make a prediction}

\item{check:}{logical indicating if dimension checks should be performed}
}
\value{
Named numeric containing c(min.prob, accuracy, freq, sr), or "no predictions" if min.prob is too high
}
\description{
Evaluate predictions performance against true label
}
