% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.adjusted.accuracy.R
\name{freq.adjusted.accuracy}
\alias{freq.adjusted.accuracy}
\title{Frequency adjusted accuracy}
\usage{
freq.adjusted.accuracy(p, q, m, n)
}
\arguments{
\item{p:}{accuracy}

\item{q:}{frequency of prediction other than hold}

\item{m:}{Average number of predictions made per day}

\item{n:}{Number of stocks}
}
\value{
: Frequency adjusted accuracy (scalar or vector)
}
\description{
Frequency adjusted accuracy
}
