% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.acc.freq.R
\name{get.acc.freq}
\alias{get.acc.freq}
\title{Get accuracy and frequency for all minimum supports}
\usage{
get.acc.freq(y, probs, granularity, set, conf)
}
\arguments{
\item{y}{numeric -1, +1}

\item{probs}{numeric indicating probability of up}

\item{granularity}{numeric indicating granularity for the sequence of min support}

\item{set}{optional character indicating the name of the set}

\item{conf}{optional numeric inidcating the confidence level to compute confidence bounds}
}
\value{
dataframe with columns min.prob, Accuracy, Frequency, Set if specified, lower and upper if conf is specified
}
\description{
Get accuracy and frequency for all minimum supports
}
