% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.exp.sr.R
\name{metric.exp.sr}
\alias{metric.exp.sr}
\title{Metric expected SR}
\usage{
metric.exp.sr(y_true, y_pred)
}
\arguments{
\item{y_true:}{tensor of shape (n obs, 1) containing true labels (1 or 0)}

\item{y_pred:}{tensor of shape (n obs, 1) containing predicted prob that obs is 1}
}
\value{
mean expected SR across min prob
}
\description{
Metric expected SR
}
