% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.mean.accuracy.R
\name{metric.mean.accuracy}
\alias{metric.mean.accuracy}
\title{Metric mean accuracy across min probs}
\usage{
metric.mean.accuracy(y_true, y_pred)
}
\arguments{
\item{y_true:}{tensor of shape (n obs, 1) containing true labels (1 or 0)}

\item{y_pred:}{tensor of shape (n obs, 1) containing predicted prob that obs is 1}
}
\value{
mean accuracy across min prob
}
\description{
Metric mean accuracy across min probs
}
