% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function
Source: http://www.guru-gis.net/multiplot-function-for-ggplot/}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL,
  title = "")
}
\arguments{
\item{...}{plots}

\item{plotlist}{list of plots}

\item{cols}{number of plots in one line}

\item{layout}{If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE), then plot 1 will go in the upper left, 2 will go in the upper right, and 3 will go all the way across the bottom.}

\item{title}{main title}
}
\value{
ggplot
}
\description{
Multiple plot function
Source: http://www.guru-gis.net/multiplot-function-for-ggplot/
}
