% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.metric.R
\name{perf.metric}
\alias{perf.metric}
\title{Performance Metric}
\usage{
perf.metric(freq.acc, metric)
}
\arguments{
\item{freq.acc}{data frame containing at least columns Freqency, Accuracy and Sr if metric is "mean.sr"}

\item{metric}{c("mean.accuracy", "mean.sr")}
}
\value{
Performance metric computed as a mean accros frequency levels
}
\description{
Performance Metric
}
