% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.freq.vs.accuracy.multiple.R
\name{plot.freq.vs.accuracy.multiple}
\alias{plot.freq.vs.accuracy.multiple}
\title{Plot frequency vs accuracy of prediction for all 3 samples}
\usage{
\method{plot}{freq.vs.accuracy.multiple}(y.train, train.probs, y.dev, dev.probs,
  y.test, test.probs, conf, stock, horizon)
}
\arguments{
\item{y.train}{numeric -1, +1}

\item{train.probs}{numeric indicating probability of up}

\item{y.dev}{numeric -1, +1}

\item{dev.probs}{numeric indicating probability of up}

\item{y.test}{numeric -1, +1}

\item{test.probs}{numeric indicating probability of up}

\item{conf}{numeric indicating if confidence intervall to be plotted}

\item{stock}{stock name for title}

\item{horizon}{horizon in minutes for title}
}
\value{
ggplot
}
\description{
Plot frequency vs accuracy of prediction for all 3 samples
}
