% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.freq.vs.accuracy.R
\name{plot.freq.vs.accuracy}
\alias{plot.freq.vs.accuracy}
\title{Plot frequency vs accuracy of prediction}
\usage{
\method{plot}{freq.vs.accuracy}(y, probs, conf, set, ytitle = T, legend = T)
}
\arguments{
\item{y}{numeric -1, +1}

\item{probs}{numeric indicating probability of up}

\item{conf}{numeric indicating if confidence intervall to be plotted (if missing, no intervall is plotted)}

\item{set}{character (used for title)}

\item{ytitle}{logical indicating if y title should be diplayed}

\item{legend}{logical indicating if y legend should be displayed}
}
\value{
ggplot
}
\description{
Plot frequency vs accuracy of prediction
}
