% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.freqacc.byset.byhoriz.R
\name{plot.freqacc.byset.byhoriz}
\alias{plot.freqacc.byset.byhoriz}
\title{Frequency vs accuracy of prediction for all 3 sets and all horizons}
\usage{
\method{plot}{freqacc.byset.byhoriz}(freq.acc, conf, stock, model)
}
\arguments{
\item{freq.acc}{data frame with columns min.prob, Accuracy, Frequency, Set, Horizon, lower, upper}

\item{conf}{numeric indicating if confidence intervall to be plotted (only for legend). !!! It must be the same as the one used to compute freq.acc, else there's a missmatch between plot and legend}

\item{stock}{stock name for title}

\item{model}{model name for title}
}
\value{
4 by 3 facets ggplot
}
\description{
Frequency vs accuracy of prediction for all 3 sets and all horizons
}
