% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.nn.R
\name{train.nn}
\alias{train.nn}
\title{Train feed-forward ANN}
\usage{
train.nn(x.train, x.dev, y.train, y.test, epochs, batch.size, activations,
  structure, l2, learning.rate)
}
\arguments{
\item{x.train}{matrix/xts n.obs*n.features}

\item{x.dev}{matrix/xts n.obs*n.features}

\item{y.train}{factor with levels c(-1, 1)}

\item{y.test}{factor with levels c(-1, 1)}

\item{epochs}{number of eposch}

\item{batch.size}{min batch size}

\item{activations}{activation functions used in hidden layer: c("relu", "tanh", ...). For the output layer, softmax is used}

\item{structure}{numeric vector indicating number of nodes in each hidden layer}

\item{l2}{l2 regularisation value}

\item{learning.rate}{}
}
\value{
fit
}
\description{
Train feed-forward ANN
}
