% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.min.prob.R
\name{best.min.prob}
\alias{best.min.prob}
\title{Find min.prob threshold maximizing dev SR}
\usage{
best.min.prob(y.dev, dev.probs)
}
\arguments{
\item{y.dev}{: factor with levels -1, 1}

\item{dev.probs}{: matrix with 2 columns: first is probability of up, second is down}
}
\value{
Named numeric containing the results of the best value for min.prop: c(min.prob, dev.accuracy, dev.freq, dev.sr)
}
\description{
Find min.prob threshold maximizing dev SR
}
