% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.forecasts.R
\name{combine.forecasts}
\alias{combine.forecasts}
\title{Combine Forecasts
Combine forecasts of various classifiers into one single forecast}
\usage{
combine.forecasts(predictions.dev, predictions.test, y.dev)
}
\arguments{
\item{predictions:}{matrix where each column contain predictions (up probability) for a model.}

\item{y:}{vector of true classes (+1 / -1)}
}
\value{
list of 2 containing vector of combined probs and fit
}
\description{
Combine Forecasts
Combine forecasts of various classifiers into one single forecast
}
