% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.indicators.R
\name{compute.indicators}
\alias{compute.indicators}
\title{Compute Indicators
Computes technical indicators based on OLHCV prices series}
\usage{
compute.indicators(prices)
}
\arguments{
\item{prices}{OLHCV prices series; xts}
}
\value{
xts with one column per technical indicator. Index is same as index of prices.
}
\description{
Compute Indicators
Computes technical indicators based on OLHCV prices series
}
