% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.lagged.ret.R
\name{compute.lagged.ret}
\alias{compute.lagged.ret}
\title{Compute lagged returns
Compute lagged return each day up to a specified lag}
\usage{
compute.lagged.ret(prices, max.lag = 120, loading = T, volume = T)
}
\arguments{
\item{prices}{OLHCV prices series; xts}

\item{max.lag}{indicates how far back lagged returns should be computed}

\item{loading}{logical telling if loading bar should be displayed}

\item{volume}{logical indicating if volume should be added}
}
\value{
xts with one column per lag. Index is same as index of prices.
}
\description{
Compute lagged returns
Compute lagged return each day up to a specified lag
}
