% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.bound.R
\name{confidence.bound}
\alias{confidence.bound}
\title{Confidence Bound}
\usage{
confidence.bound(p, n, conf = 0.95)
}
\arguments{
\item{p}{Sample proportion; numeric}

\item{n}{Number of observations; numeric}

\item{conf:}{confidence level within the bounds}
}
\value{
dataframe with values of both bounds of each element
}
\description{
Confidence Bound
}
\examples{
confidence.bound(c(0.7, 0.8), c(100000, 500), 0.95)
}
