% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.vs.accuracy.multiplot.R
\name{freq.vs.accuracy.multiplot}
\alias{freq.vs.accuracy.multiplot}
\title{Frequency vs accuracy of prediction for all 3 samples}
\usage{
freq.vs.accuracy.multiplot(y.train, train.probs, y.dev, dev.probs, y.test,
  test.probs, conf, stock, horizon, granularity = 0.01)
}
\arguments{
\item{y.train}{numeric -1, +1}

\item{train.probs}{numeric indicating probability of up}

\item{y.dev}{numeric -1, +1}

\item{dev.probs}{numeric indicating probability of up}

\item{y.test}{numeric -1, +1}

\item{test.probs}{numeric indicating probability of up}

\item{conf}{numeric indicating if confidence intervall to be plotted}

\item{stock}{stock name for title}

\item{horizon}{horizon in minutes for title}

\item{granularity}{numeric indicating granularity for the sequence of min support}
}
\value{
ggplot
}
\description{
Frequency vs accuracy of prediction for all 3 samples
}
