% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.acc.freq.R
\name{get.acc.freq}
\alias{get.acc.freq}
\title{Get accuracy and frequency for all minimum supports}
\usage{
get.acc.freq(y, probs, granularity, set, horizon, conf, min.preds = 0,
  sr = F, n = 7)
}
\arguments{
\item{y}{numeric -1, +1}

\item{probs}{numeric indicating probability of up}

\item{granularity}{numeric indicating granularity for the sequence of min support}

\item{set}{optional character indicating the name of the set}

\item{conf}{optional numeric inidcating the confidence level to compute confidence bounds}

\item{min.preds}{numeric indicating the minimum number of predictions for a frequency level to be considered}

\item{sr}{logical indicating if sr should be returned}

\item{Horizon}{optional numeric indicating the horizon}
}
\value{
dataframe with columns min.prob, Accuracy, Frequency, Set if specified, Horizon if specified, lower and upper if conf is specified
}
\description{
Get accuracy and frequency for all minimum supports
}
