% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.freq.vs.accuracy.R
\name{plot.freq.vs.accuracy}
\alias{plot.freq.vs.accuracy}
\title{Plot frequency vs accuracy of prediction}
\usage{
\method{plot}{freq.vs.accuracy}(y, probs, conf, stock = "stock", horizon)
}
\arguments{
\item{y}{: factor with levels -1, 1}

\item{probs}{: matrix with 2 columns: first is probability of up, second is down}

\item{conf:}{numeric indicating if confidence intervall to be plotted (if missing, no intervall is plotted)}

\item{stock:}{character (used for title)}

\item{horizon:}{character or numeric (used for title)}
}
\value{
ggplot
}
\description{
Plot frequency vs accuracy of prediction
}
